+++ src/features/basic/index.ts
@@ -43,6 +43,7 @@ import './prevent-delete-button-misclicks';
 import './prod-burn-link';
 import './prod-order-eta';
 import './prodq-queue-load';
+import './production-order-error-highlight';
 import './prun-bugs';
 import './rprun-version-label';
 import './screen-tab-bar/screen-tab-bar';